backend/requirements.txt
fastapi
uvicorn
pydantic
google-genai
python-dotenv
PyGithub
pinecone
